requirements.txt
streamlit
pandas
numpy
plotly
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fpdf
python-dotenv
scikit-learn
